Move session entity to sessions folder
