+++ backend/src/auth/auth.service.spec.ts
@@ -18,7 +18,7 @@ import {
 } from '../errors/errors';
 import { Identity } from '../identity/identity.entity';
 import { LoggerModule } from '../logger/logger.module';
-import { Session } from '../users/session.entity';
+import { Session } from '../sessions/session.entity';
 import { User } from '../users/user.entity';
 import { UsersModule } from '../users/users.module';
 import { AuthToken } from './auth-token.entity';
