+++ backend/src/realtime/websocket/websocket.gateway.spec.ts
@@ -35,9 +35,9 @@ import { PermissionsModule } from '../../permissions/permissions.module';
 import { PermissionsService } from '../../permissions/permissions.service';
 import { Edit } from '../../revisions/edit.entity';
 import { Revision } from '../../revisions/revision.entity';
-import { SessionModule } from '../../session/session.module';
-import { SessionService } from '../../session/session.service';
-import { Session } from '../../users/session.entity';
+import { Session } from '../../sessions/session.entity';
+import { SessionModule } from '../../sessions/session.module';
+import { SessionService } from '../../sessions/session.service';
 import { User } from '../../users/user.entity';
 import { UsersModule } from '../../users/users.module';
 import { UsersService } from '../../users/users.service';
