+++ backend/src/realtime/websocket/websocket.gateway.ts
@@ -16,7 +16,7 @@ import { ConsoleLoggerService } from '../../logger/console-logger.service';
 import { NotesService } from '../../notes/notes.service';
 import { NotePermission } from '../../permissions/note-permission.enum';
 import { PermissionsService } from '../../permissions/permissions.service';
-import { SessionService } from '../../session/session.service';
+import { SessionService } from '../../sessions/session.service';
 import { User } from '../../users/user.entity';
 import { UsersService } from '../../users/users.service';
 import { RealtimeConnection } from '../realtime-note/realtime-connection';
