+++ backend/src/seed.ts
@@ -19,7 +19,7 @@ import { NoteGroupPermission } from './permissions/note-group-permission.entity'
 import { NoteUserPermission } from './permissions/note-user-permission.entity';
 import { Edit } from './revisions/edit.entity';
 import { Revision } from './revisions/revision.entity';
-import { Session } from './users/session.entity';
+import { Session } from './sessions/session.entity';
 import { User } from './users/user.entity';
 import { hashPassword } from './utils/password';
