+++ backend/src/sessions/session.entity.ts
