+++ backend/src/sessions/session.module.ts
@@ -6,7 +6,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { Session } from '../users/session.entity';
+import { Session } from './session.entity';
 import { SessionService } from './session.service';
 
 @Module({
