+++ backend/test/test-setup.ts
@@ -77,8 +77,8 @@ import { PermissionsModule } from '../src/permissions/permissions.module';
 import { PermissionsService } from '../src/permissions/permissions.service';
 import { RevisionsModule } from '../src/revisions/revisions.module';
 import { RevisionsService } from '../src/revisions/revisions.service';
-import { SessionModule } from '../src/session/session.module';
-import { SessionService } from '../src/session/session.service';
+import { SessionModule } from '../src/sessions/session.module';
+import { SessionService } from '../src/sessions/session.service';
 import { User } from '../src/users/user.entity';
 import { UsersModule } from '../src/users/users.module';
 import { UsersService } from '../src/users/users.service';
